EPMRPP-91636 || move search field to fields
