+++ app/src/components/fields/searchField/searchField.scss
